Support an instance mode for node-html-to-image
**This is a breaking API change (for now)**

node-html-to-image's default export is now a function which returns an
instance object with a render() method which is the old
nodeHtmlToImage() functionality. The reason for doing this is to allow
us to instantiate a single 'image maker' object that we call render() on
multiple times and reuse the same puppeteer cluster.

In addition,
* Update the type definitions.
* Use Cluster.execute() instead of Cluster.queue() plus Cluster.idle()
  as this will allow better use in concurrent situations. We will only
  wait for our jobs to finish rather than for the entire cluster to go
  idle.

Fixes frinyvonnick/node-html-to-image#80